Fix all type conversion warnings in Phase 66-70 files
Fixed type conversion warnings in:
- simulation/include/simulation/timestepping/AdaptiveTimestepper.h
  * Line 226: Added explicit cast in acceptance rate calculation
  * Line 288: Cast size_t to double in relative error norm
  * Line 318: Cast size_t to double in scaled error calculation

- simulation/include/simulation/stability/StabilityMonitor.h
  * Lines 329-330: Changed int to size_t for sign_changes counters
  * Line 333: Changed int to size_t for loop variable
  * Line 346: Added explicit cast in oscillation index return

- simulation/include/simulation/coupling/ThermalChemicalCoupling.h
  * Line 70: Cast int to size_t in vector resize
  * Line 72: Cast int to size_t in array access
  * Line 125: Cast int to size_t for product_ids array access
  * Line 132: Cast int to size_t for reactant_ids array access
  * Line 305: Mark unused parameter with [[maybe_unused]]

- examples/full_simulation_example.cpp
  * Lines 64-67: Cast int to size_t in vector constructors
  * Lines 107-111: Cast int to size_t in array indexing
  * Lines 233-242: Use size_t index in diffusion loop
  * Lines 245-260: Use size_t index in reaction loop
  * Line 293: Mark argc/argv as [[maybe_unused]]

All Phase 66-70 files now compile cleanly with -Werror

+++ simulation/include/simulation/stability/StabilityMonitor.h
@@ -326,11 +326,11 @@ private:
 
         // Count sign changes in derivative
         size_t n_vars = solution_history_[0].size();
-        int total_sign_changes = 0;
-        int max_possible_changes = (solution_history_.size() - 2) * n_vars;
+        size_t total_sign_changes = 0;
+        size_t max_possible_changes = (solution_history_.size() - 2) * n_vars;
 
         for (size_t i = 0; i < n_vars; ++i) {
-            int sign_changes = 0;
+            size_t sign_changes = 0;
             for (size_t t = 1; t < solution_history_.size() - 1; ++t) {
                 double deriv_prev = solution_history_[t][i] - solution_history_[t-1][i];
                 double deriv_next = solution_history_[t+1][i] - solution_history_[t][i];
@@ -343,7 +343,7 @@ private:
         }
 
         // Oscillation index: fraction of sign changes
-        return static_cast<double>(total_sign_changes) / max_possible_changes;
+        return static_cast<double>(total_sign_changes) / static_cast<double>(max_possible_changes);
     }
 };
